Added Search Bar-- Search bar filters the table and shows whatever coin you search for! Noticed that landscape view is broken beyond belief. Will try to fix in the next commit.
Also adding several "unversioned files" to the git... Not sure what those are doing not being committed.
This includes the Gson library, and literally everything to do with Gradle.

Still need to prune the BTC- , ETH- , USDT- from all of the cashtags. Should be a simple method in Currency.java.
Still want to look at the "holder" for the adapter.
Need to do a little bit of refactoring. (customAdapter -> CustomAdapter)

And of course, Story ID 4 needs to be worked on.

Story ID 1 and 3 complete. (Table and Search Bar)

(Story ID 1 was actually complete with the inital table commit but forgot to mention it)