Added Favoriting and Banner--
A huge update! Very exciting.

Clicking on a coin's name now favorites the coin, although because no sorting is done yet, this is a purely cosmetic change for the coin's background.
However, what ISN'T cosmetic is the addition of shared preferences! We don't have to be connected to the internet to use the app anymore!
And on top of that, a HUGE change is the scrolling banner for USDT prices. Added it for fun because it looks cool.

Also did some refactoring. Documentation is okay right now too.

Story ID 4 complete-- Sprint 1 technically complete!
But I still want to add sorting...
We'll see tomorrow.