Sprint 2 Commit 1 Sorting Buttons--
Sortable buttons! Sorting by name, price, and if the coin is favorited or not.
Currently shoves all USDT markets at the bottom of the list.

Currency class now contains many Comparators for the sorting mentioned above.

Added some checking for if there is any internet before the AsynchTask for fetching data is run.

The Currencies ArrayList is now no longer guaranteed to be sorted as the order is now directly modified by the sorting functions.

A slight reorganization of code may be on the horizon, but everything looks reasonable so far.

REMOVED SUPPORT FOR ETHEREUM MARKETS (Got too complicated with the sorting)

BUGS:
---------
When closing the app after favoriting when querying, the list of currencies is deleted entirely and all other favorites are lost.

Possible solutions:
1. Bind the adapter to a copy of the currencies and only UPDATE in the main list when saving.
2. Experiment with "onStop()" method and have it save the entire list in an expected way whenever the app closes. (Maybe best option)
---------
When sorting by favorites, the USDT markets are thrown into the middle of the list

Solution:
Add extra logic to the favorites comparator (lengthy process)
---------