Sprint 3 Commit 1- Basic Charts (Axes are messed up)--
Favoriting now is done by clicking anywhere on the coin but the name.

Clicking on names (which are currently a temporary blue color to give the appearance of a link) now displays in a separate activity...
...
...
A CHART!

This chart is very basic. The Axes on the bottom are unlabelled until I find a way to properly label them, and currently the chart can only display hour ticks and you cannot highlight values to see exact information. Zooming is untested as I did not have an android phone to test on today or yesterday. Testing WILL be done with zoom soon, but I wasn't able to get to it.

 There was also a major refactor with the AsyncTask (GetFeed)
 This is now a much more generalized object and all activities that use it now should implement AsyncResponse to use GetFeed.
 The GetFeed constructor now accepts a class that has implemented AsyncResponse as well.
 The Custom Adapter also changed quite a bit, as it has to pass data into the SharedPreferences file to pass data to the chart activity.

 A runthrough should almost certainly be done to update documentation.

 The favorited list bug with the USDT markets being in the middle has not been addressed.