Sprint 3 Commit 4 Charts Different Intervals & Start of major refactoring--
Story ID 2 Complete-(ish)

Charts now have different interval options selected by radio buttons. (This took forever) I'm going to call Story ID 2 tentatively complete because testing hasn't really been rigorous as of yet, but I will note an update if something breaks later and I have to fix it.

This also marked the first refactor: There is now a database wrapper class whose purpose is to be the sole entity that interacts with the database.
Next up is the General class which will be a class for all sorts of functions that multiple classes have uses for.

The next commit should have taken out the favoriting logic from the custom adapter, as the logic should not be tied into the adapter.