Sprint 3 Commit 4 (Actual Commit 4) Final Commit for Sprint 3--
First, there was a typo that I tried and failed to fix in the title of the last commit. It should have been commit 3, not commit 4.

Fixed the sort by favorites bug, had a slight refactor on networking.

Did not get to removing the favoriting logic from the custom adapter or decoupling the sorting buttons from the set adapter function.