Sprint 4 Commit 1 Highlighting Values in Chart--
First, a major refactor was done and many major parts of the main activity were split up. The sorting buttons were also finally uncoupled from the list adapter function and other things were made better organized.

However, the main thing here is the highlighting of values in the charts! If a candle is clicked on now, it will show specific data for the candle in the area above the chart.