Sprint 4 Commit 2 Volume in Chart and Toggle Volume Button--
Volume can now be viewed in the chart. It can be toggled on and off with the button below the chart.
Currently, Volume uses the left axis to show its data and currently draws lines from the left axis as well.

As per the new norm, there have been a couple of refactors with this commit as well. Namely with the drawChart function on the ChartActivity.